+++ src/libmozok/libmozok/quest.cpp
@@ -1,19 +1,47 @@
 // Copyright 2024 Pavlo Savchuk. Subject to the MIT license.
 
+#include <libmozok/public_types.hpp>
+#include <libmozok/statement.hpp>
 #include <libmozok/quest.hpp>
 
 namespace mozok {
 
 namespace {
-    const ActionPtr nullAction(nullptr);
-}
+
+const ActionPtr nullAction(nullptr);
+const StatePtr nullState(nullptr);
+
+/// @brief Fills the `PossibleActionVec` vector with data provided by an 
+///        `iterateOverApplicableActions` call.
+class PossibleActionsBuilder : public QuestApplicableActionsIterator {
+    Quest::PossibleActionVec& _possibleActions;
+public:
+    PossibleActionsBuilder(
+            Quest::PossibleActionVec& possibleActions)
+        : _possibleActions(possibleActions)
+    { /* empty */ }
+
+    virtual bool actionCallback(
+            const ActionPtr& action, 
+            const ObjectVec& arguments,
+            const SIZE_T combinedIndx
+            ) noexcept {
+        _possibleActions.push_back({action, arguments, combinedIndx});
+        return true;
+    }
+};
+
+} // namespace
 
 
 QuestApplicableActionsIterator::~QuestApplicableActionsIterator() noexcept 
     = default;
     
 bool QuestApplicableActionsIterator::actionCallback(
-        const ActionPtr& /*action*/, const ObjectVec& /*arguments*/) noexcept
+        const ActionPtr& /*action*/, 
+        const ObjectVec& /*arguments*/,
+        const SIZE_T /*combinedIndx*/
+    ) noexcept
 { return true; }
 
 
@@ -37,7 +65,8 @@ Quest::Quest(
     _idToAction(buildIdToActionMap()),
     _relevantActions(buildRelevantActions(actions)),
     _relevantObjects(buildRelevantObjects(objects)),
-    _relevantRelations(buildRelevantRelations(actions))
+    _relevantRelations(buildRelevantRelations(actions)),
+    _possibleActions(buildPossibleActions())
 { /* empty */ }
 
 Quest::ActionArgObjects Quest::buildActionArgObjects() const noexcept {
@@ -55,7 +84,7 @@ Quest::ActionArgObjects Quest::buildActionArgObjects() const noexcept {
             if(argObjects.size() == 0) {
                 notEmpty = false;
                 break;
-            }
+}
         }
         if(notEmpty) {
             // Action can be applicable.
@@ -110,6 +139,14 @@ UnorderedSet<ID> Quest::buildRelevantRelations(
     return res;
 }
 
+Quest::PossibleActionVec Quest::buildPossibleActions() const noexcept {
+    PossibleActionVec possibleActions;
+    PossibleActionsBuilder it(possibleActions);
+    Vector<StatementVec> emptyPre;
+    iterateOverApplicableActions_Slow(nullState, it, emptyPre);
+    return possibleActions;
+}
+
 const Str& Quest::getName() const noexcept {
     return _name;
 }
@@ -145,7 +182,11 @@ const QuestVec& Quest::getSubquests() const noexcept {
     return _subquests;
 }
 
-void Quest::iterateOverApplicableActions(
+const Quest::PossibleActionVec& Quest::getPossibleActions() const noexcept {
+    return _possibleActions;
+}
+
+void Quest::iterateOverApplicableActions_Slow(
             const StatePtr& state,
             QuestApplicableActionsIterator& it,
             Vector<StatementVec>& actionPreBuffers
@@ -157,10 +198,31 @@ void Quest::iterateOverApplicableActions(
             if(action->getArguments().size() > 0)
                 continue; // This action is not applicable.
         ObjectVec objects(action->getArguments().size(), ObjectPtr(nullptr));
-        ObjectSet objectSet;
         if(findNextObj(state, it, actionPreBuffers, 
-                objects, objectSet, actionIndx, 0) == false)
+                objects, actionIndx, 0, SIZE_T(0), SIZE_T(1)) == false)
             break; // Stop the search.
+    }
+}
+
+void Quest::iterateOverApplicableActions(
+            const StatePtr& state,
+            QuestApplicableActionsIterator& it,
+            Vector<StatementVec>& actionPreBuffers
+            ) const noexcept {
+    for(const ActionWithArgs& aa : _possibleActions) {
+        // Objects were selected in such a way that they are suitable by types,
+        // but we need to verify if they also satisfy the action preconditions.
+        if(state != nullState) {
+            if(aa.action->checkActionPreconditions(
+                    aa.arguments, state, 
+                    actionPreBuffers[aa.combinedIndx % _actions.size()]
+                    ) == false)
+                continue;
+        }
+        if(it.actionCallback(
+                aa.action, aa.arguments, 
+                aa.combinedIndx) == false)
+            break;
     }
 }
 
@@ -169,31 +231,43 @@ bool Quest::findNextObj(
         QuestApplicableActionsIterator& it,
         Vector<StatementVec>& actionPreBuffers,
         ObjectVec &objects,
-        ObjectSet &objectSet,
         ObjectVec::size_type actionIndx,
-        ObjectVec::size_type argIndx
+        ObjectVec::size_type argIndx,
+        SIZE_T combinedIndx,
+        SIZE_T combinedSize
         ) const noexcept {
+
     if(argIndx >= _actionArgObjects[actionIndx].size()) {
         // A new possible substitution has been found.
         const ActionPtr& action = _actions[actionIndx];
         // Objects were selected in such a way that they are suitable by types,
         // but we need to verify if they also satisfy the action preconditions.
-        if(action->checkActionPreconditions(
-                objects, state, actionPreBuffers[actionIndx]))
-            return it.actionCallback(action, objects);
-        return true;
+        if(state != nullState)
+            if(action->checkActionPreconditions(
+                    objects, state, actionPreBuffers[actionIndx]) == false)
+                return true;
+        // Call the callback function and return the call's result.
+        return it.actionCallback(
+                action, objects, 
+                actionIndx + combinedIndx * _actions.size());
     }
-    for(const ObjectPtr& obj : _actionArgObjects[actionIndx][argIndx]) {
-        if(objectSet.find(obj) != objectSet.end())
+    const SIZE_T multiplier = _actionArgObjects[actionIndx][argIndx].size();
+    for(SIZE_T i = 0; i < multiplier; ++i) {
+        const ObjectPtr& obj = _actionArgObjects[actionIndx][argIndx][i];
+        bool alreadyInUse = false;
+        for(SIZE_T j=0; j<argIndx; ++j)
+            if(objects[j]->getId() == obj->getId()) {
+                alreadyInUse = true;
+                break;
+            }
+        if(alreadyInUse)
             continue;
-        objectSet.insert(obj);
         objects[argIndx] = obj;
         if(findNextObj(
-                state, it, actionPreBuffers,
-                objects, objectSet, actionIndx, 
-                argIndx + 1) == false)
+                state, it, actionPreBuffers, objects, actionIndx, 
+                argIndx + 1, combinedIndx + i * combinedSize,
+                combinedSize * multiplier) == false)
             return false;
-        objectSet.erase(obj);
     }
     return true;
 }