+++ src/libmozok/libmozok/quest.hpp
@@ -27,11 +27,14 @@ public:
     /// @brief This method will be invoked for the next applicable action.
     /// @param action Action that is applicable.
     /// @param arguments Applicable action arguments.
+    /// @param combinedIndex Unique combined index of action and arguments.
+    /// This is always in the form `actionIndex + actionCount * argumentIndex`.
     /// @return Return `true` if you want to continue the search. 
     ///         Return `false` if you want to stop the search.
     virtual bool actionCallback(
             const ActionPtr& /*action*/, 
-            const ObjectVec& /*arguments*/
+            const ObjectVec& /*arguments*/,
+            const SIZE_T /*combinedIndx*/
             ) noexcept;
 };
 
@@ -85,20 +88,54 @@ class Quest {
     /// @brief The set of relevant relation IDs.
     const UnorderedSet<ID> _relevantRelations;
     
+public:
+    struct ActionWithArgs {
+        ActionPtr action; 
+        ObjectVec arguments;
+        SIZE_T combinedIndx;
+    };
+    using PossibleActionVec = Vector<ActionWithArgs>;
+    const PossibleActionVec& getPossibleActions() const noexcept;
+
+private:
+    /// @brief Provides all possible actions (with their respective arguments)
+    ///        that can be executed.
+    const PossibleActionVec _possibleActions;
+
+
     UnorderedSet<ID> buildRelevantActions(const ActionVec& actions) const noexcept;
     UnorderedSet<ID> buildRelevantObjects(const ObjectVec& objects) const noexcept;
     UnorderedSet<ID> buildRelevantRelations(
             const ActionVec& actions) const noexcept;
+    PossibleActionVec buildPossibleActions() const noexcept;
     
-    /// @brief Iterates trough all allowed objects for a given allowed action.
-    /// @param state The state from which the search occurs.
+    /// @brief Iterates through all potential applicable actions. 
+    ///         This version does not use pre-calculated `_possibleActions`.
+    /// @param state The state from which do the search. If `state` is 
+    ///         `nullptr`, it will not check the action preconditions.
     /// @param it Callback object.
     /// @param actionPreBuffers Action's pre-buffer (see 
     ///         `QuestPlanner::_actionPreBuffers` for the description).
+    ///         Not used when state is `nullptr`.
+    void iterateOverApplicableActions_Slow(
+            const StatePtr& state,
+            QuestApplicableActionsIterator& it,
+            Vector<StatementVec>& actionPreBuffers
+            ) const noexcept;
+
+    /// @brief Iterates trough all allowed objects for a given allowed action.
+    /// @param state The state from which the search occurs. If `state` is 
+    ///         `nullptr` then it will skip checking the action preconditions.
+    /// @param it Callback object.
+    /// @param actionPreBuffers Action's pre-buffer (see 
+    ///         `QuestPlanner::_actionPreBuffers` for the description).
+    ///         Not used when state is `nullptr`.
     /// @param objects Current list of selected allowed objects.
-    /// @param objectSet The set of all currently selected allowed objects.
     /// @param actionIndx Action's index in the list of quest's allowed actions.
     /// @param argIndx Action argument index (starting from 0).
+    /// @param combinedIndx Current combined index of all selected arguments.
+    /// @param combinedSize Current combined size of all possible first 
+    ///         `argIndx` arguments.
     /// @return Returns true if no substitutions were found. Returns false if at 
     ///         some point the callback object halts the search. Otherwise, 
     ///         returns true.
@@ -107,9 +144,10 @@ class Quest {
             QuestApplicableActionsIterator& it,
             Vector<StatementVec>& actionPreBuffers,
             ObjectVec &objects,
-            ObjectSet &objectSet,
             ObjectVec::size_type actionIndx,
-            ObjectVec::size_type argIndx
+            ObjectVec::size_type argIndx,
+            SIZE_T combinedIndx,
+            SIZE_T combinedSize
             ) const noexcept;
 
 public:
@@ -133,10 +171,12 @@ public:
     const QuestVec& getSubquests() const noexcept;
 
     /// @brief Iterates trough the possible applicable actions.
-    /// @param state The state from which the search occurs.
+    /// @param state The state from which the search occurs. If `state` is 
+    ///         `nullptr`, it will not check the action preconditions.
     /// @param it Callback object.
     /// @param actionPreBuffers Action's pre-buffer (see 
     ///         `QuestPlanner::_actionPreBuffers` for the description).
+    ///         Not used when state is `nullptr`.
     void iterateOverApplicableActions(
             const StatePtr& state,
             QuestApplicableActionsIterator& it,